# web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AIAnalysisDetail": {
        "additionalProperties": false,
        "description": "The stored per-trade AI review, as the page reads it.\n\nDeliberately NOT the row. `cost_usd`, `tokens_input`, `tokens_output`,\nevery `*_job_id` and `raw_response_json` are absent: the raw response is\nunvalidated model output and cost is billing detail, and neither belongs\nin a browser.\n\n`confirmed_fields` travels because the panel has to distinguish a label\nthe AI produced from one the trader stands behind \u2014 that difference is\nthe whole point of the lock, and a UI that cannot see it would present\nthe trader's own decision as a machine's guess.\n\n`latest_proposals` carries what the NEWEST analysis read for each label,\nincluding the ones the trader has locked. Design decision 3: locked\nmeans \"not applied\", never \"hidden\". Without it a trader can never see\nwhat a re-analysis found for a confirmed field \u2014 they would have to\nrelease it, re-run, and hope. This is a projection through the SAME\nconfirmable allowlist, not the raw model output: `raw_response_json`\nstays server-side, and every value here is one the trader could already\nhave set by hand.",
        "properties": {
          "ai_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Grade"
          },
          "bias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bias"
          },
          "confirmed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Confirmed Fields",
            "type": "array"
          },
          "detected_setup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Setup"
          },
          "grading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIGrading"
              },
              {
                "type": "null"
              }
            ]
          },
          "journal_entry_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journal Entry Md"
          },
          "key_zones": {
            "items": {
              "type": "string"
            },
            "title": "Key Zones",
            "type": "array"
          },
          "latest_proposals": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Latest Proposals",
            "type": "object"
          },
          "matched_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Strategy"
          },
          "missed_opportunities": {
            "items": {
              "type": "string"
            },
            "title": "Missed Opportunities",
            "type": "array"
          },
          "possible_mistakes": {
            "items": {
              "type": "string"
            },
            "title": "Possible Mistakes",
            "type": "array"
          },
          "trade_quality": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Quality"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Grade"
          }
        },
        "required": [
          "bias",
          "detected_setup",
          "trade_quality",
          "matched_strategy",
          "key_zones",
          "possible_mistakes",
          "missed_opportunities",
          "journal_entry_md",
          "grading",
          "ai_grade",
          "user_grade",
          "confirmed_fields",
          "latest_proposals",
          "updated_at"
        ],
        "title": "AIAnalysisDetail",
        "type": "object"
      },
      "AIAnalysisJobRequest": {
        "additionalProperties": false,
        "description": "Which of the caller's own screenshots to analyse for this trade.\n\nA screenshot id, not a key and not a URL: the bytes analysed are the\npromoted object `finalize_upload` produced, and this is the only handle\nthe browser has on one. Ownership is never input.",
        "properties": {
          "screenshot_id": {
            "title": "Screenshot Id",
            "type": "integer"
          }
        },
        "required": [
          "screenshot_id"
        ],
        "title": "AIAnalysisJobRequest",
        "type": "object"
      },
      "AIAnalysisLabelPatch": {
        "additionalProperties": false,
        "description": "The labels a trader may confirm or correct. A positive allowlist.\n\nServer-owned columns \u2014 `cost_usd`, `tokens_input`, `raw_response_json`,\nevery `*_job_id`, `confirmed_at` \u2014 are absent by construction, and\n`extra=\"forbid\"` turns sending one into a 422 rather than a silent drop.\n\n`release` hands named labels back to the AI. It is the only way out of\nthe confirmation lock, and it is an allowlist for the same reason the\nvalues are: it must never become a way to name arbitrary columns.",
        "properties": {
          "bias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bias"
          },
          "detected_setup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Setup"
          },
          "matched_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Strategy"
          },
          "release": {
            "default": [],
            "items": {
              "enum": [
                "bias",
                "detected_setup",
                "trade_quality",
                "matched_strategy"
              ],
              "type": "string"
            },
            "title": "Release",
            "type": "array"
          },
          "trade_quality": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Quality"
          },
          "user_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Grade"
          }
        },
        "title": "AIAnalysisLabelPatch",
        "type": "object"
      },
      "AIAnalysisLabels": {
        "additionalProperties": false,
        "properties": {
          "bias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bias"
          },
          "confirmed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Confirmed Fields",
            "type": "array"
          },
          "detected_setup": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Setup"
          },
          "matched_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Strategy"
          },
          "trade_quality": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Quality"
          },
          "user_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Grade"
          }
        },
        "required": [
          "bias",
          "detected_setup",
          "trade_quality",
          "matched_strategy",
          "user_grade",
          "confirmed_fields"
        ],
        "title": "AIAnalysisLabels",
        "type": "object"
      },
      "AIGrading": {
        "additionalProperties": false,
        "properties": {
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "one_line_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "One Line Verdict"
          },
          "rubric": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AIGradingRubricEntry"
            },
            "title": "Rubric",
            "type": "object"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "grade",
          "score",
          "one_line_verdict",
          "rubric"
        ],
        "title": "AIGrading",
        "type": "object"
      },
      "AIGradingRubricEntry": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "score",
          "note"
        ],
        "title": "AIGradingRubricEntry",
        "type": "object"
      },
      "AIJobAccepted": {
        "additionalProperties": false,
        "properties": {
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "created"
        ],
        "title": "AIJobAccepted",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TradeLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/trades/{trade_id}/analysis": {
      "get": {
        "description": "The stored AI review for one of the caller's own trades.\n\nA trade with no analysis yet is a 404, not an empty object: \"not run\"\nand \"run, and it found nothing\" are different states and the panel\nrenders them differently. Foreign and missing are the same 404.",
        "operationId": "get_trade_analysis_v1_trades__trade_id__analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trade Analysis",
        "tags": [
          "trades"
        ]
      },
      "patch": {
        "description": "Confirm, correct or release the AI's labels for one of the caller's trades.\n\nThe trader's judgement is the point of this route: it is what\npersonalization learns from, and it is what locks a label against every\nlater analysis job until they release it.\n\n`user_grade` is handled apart from the labels because it lives on\n`trades`, not on `aianalysis`, and must never be written by the grading\njob \u2014 that column is the trader's own verdict. `None` is a meaningful\nvalue there (it clears the override), so \"not sent\" needs its own\nsentinel rather than being inferred from a null.",
        "operationId": "patch_trade_analysis_v1_trades__trade_id__analysis_patch",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalysisLabelPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisLabels"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Trade Analysis",
        "tags": [
          "trades"
        ]
      },
      "post": {
        "description": "Queue AI analysis of one of the caller's own screenshots for one trade.\n\nOwnership of BOTH the trade and the screenshot is settled first, before\nanything is written and before any billable work is scheduled: a queued\njob is spend and, on a poll, an existence oracle. Foreign and missing are\nthe same 404.\n\nThe job reads the PROMOTED object, not an upload: `finalize_upload` has\nalready decoded, capped and re-encoded those bytes, so the model only\never sees bytes we produced.",
        "operationId": "enqueue_trade_analysis_v1_trades__trade_id__analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "trade_id",
            "required": true,
            "schema": {
              "title": "Trade Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalysisJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Trade Analysis",
        "tags": [
          "trades"
        ]
      }
    },
...
  }
}